requirements.txt
setuptools == 69.5.1
zenml == 0.57.1 
zenml[server]
python-dotenv == 1.0.1
langchain == 0.1.20
langchain-community == 0.0.38
langchain-experimental == 0.0.58
sentence-transformers == 2.7.0
pypdf == 4.2.0
rapidocr-onnxruntime == 1.3.19
langchain-google-genai == 1.0.4